+++ examples/create-react-app/src/api/FeatureApi.ts
@@ -1,7 +1,7 @@
 import { Feature, FeatureCreate, FeatureUpdate } from "../dto/Feature";
 import { ConnectorCrud } from "@jfront/core-rest";
 
-class ExampleApi extends ConnectorCrud<Feature, FeatureCreate, FeatureUpdate> {
+class ExampleApi extends ConnectorCrud<Feature, string, FeatureCreate, FeatureUpdate> {
   constructor(baseUrl: string) {
     baseUrl = baseUrl + "/feature";
     super(baseUrl, false);