+++ packages/rest/src/index.ts
@@ -3,3 +3,4 @@ export * from "./api/types";
 export * from "./api/ConnectorBase";
 export * from "./api/ConnectorCrud";
 export * from "./api/ConnectorSearch";
+export * from "./api/ConnectorChield";